Re-organize C++ torch::nn modules
