+++ test/cpp/api/functional.cpp
@@ -1,0 +1,65 @@
+#include <gtest/gtest.h>
+
+#include <torch/torch.h>
+
+#include <test/cpp/api/support.h>
+
+namespace F = torch::nn::functional;
+
+using namespace torch::nn;
+
+struct FunctionalTest : torch::test::SeedingFixture {};
+
+TEST_F(FunctionalTest, MaxPool1d) {
+  auto x = torch::ones({1, 1, 5}, torch::requires_grad());
+  auto y = F::max_pool1d(x, MaxPool1dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 3);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({1, 1 ,2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({1, 1, 2}));
+}
+
+TEST_F(FunctionalTest, MaxPool2d) {
+  auto x = torch::ones({2, 5, 5}, torch::requires_grad());
+  auto y = F::max_pool2d(x, MaxPool2dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 3);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({2, 2 ,2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({2, 2, 2}));
+}
+
+TEST_F(FunctionalTest, MaxPool3d) {
+  auto x = torch::ones({2, 5, 5, 5}, torch::requires_grad());
+  auto y = F::max_pool3d(x, MaxPool3dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 4);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({2, 2, 2, 2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({2, 2, 2, 2}));
+}
+
+TEST_F(FunctionalTest, AvgPool1d) {
+  auto x = torch::ones({1, 1, 5}, torch::requires_grad());
+  auto y = F::avg_pool1d(x, AvgPool1dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 3);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({1, 1, 2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({1, 1, 2}));
+}
+
+TEST_F(FunctionalTest, AvgPool2d) {
+  auto x = torch::ones({2, 5, 5}, torch::requires_grad());
+  auto y = F::avg_pool2d(x, AvgPool2dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 3);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({2, 2, 2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({2, 2, 2}));
+}
+
+TEST_F(FunctionalTest, AvgPool3d) {
+  auto x = torch::ones({2, 5, 5, 5}, torch::requires_grad());
+  auto y = F::avg_pool3d(x, AvgPool3dOptions(3).stride(2));
+
+  ASSERT_EQ(y.ndimension(), 4);
+  ASSERT_TRUE(torch::allclose(y, torch::ones({2, 2, 2, 2})));
+  ASSERT_EQ(y.sizes(), torch::IntArrayRef({2, 2, 2, 2}));
+}
