+++ tools/build_variables.py
@@ -186,7 +186,6 @@ def add_torch_libs():
         "torch/csrc/api/src/jit.cpp",
         "torch/csrc/api/src/nn/init.cpp",
         "torch/csrc/api/src/nn/module.cpp",
-        "torch/csrc/api/src/nn/modules/avgpool.cpp",
         "torch/csrc/api/src/nn/modules/batchnorm.cpp",
         "torch/csrc/api/src/nn/modules/conv.cpp",
         "torch/csrc/api/src/nn/modules/dropout.cpp",
@@ -195,8 +194,8 @@ def add_torch_libs():
         "torch/csrc/api/src/nn/modules/functional.cpp",
         "torch/csrc/api/src/nn/modules/linear.cpp",
         "torch/csrc/api/src/nn/modules/loss.cpp",
-        "torch/csrc/api/src/nn/modules/maxpool.cpp",
         "torch/csrc/api/src/nn/modules/named_any.cpp",
+        "torch/csrc/api/src/nn/modules/pooling.cpp",
         "torch/csrc/api/src/nn/modules/rnn.cpp",
         "torch/csrc/api/src/optim/adagrad.cpp",
         "torch/csrc/api/src/optim/adam.cpp",
