+++ torch/csrc/api/include/torch/nn.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include <torch/nn/cloneable.h>
+#include <torch/nn/functional.h>
 #include <torch/nn/init.h>
 #include <torch/nn/module.h>
 #include <torch/nn/modules.h>
+#include <torch/nn/options.h>
 #include <torch/nn/pimpl.h>
