+++ torch/csrc/api/include/torch/nn/functional.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <torch/nn/functional/pooling.h>
