+++ torch/csrc/api/include/torch/nn/modules.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <torch/nn/modules/any.h>
-#include <torch/nn/modules/avgpool.h>
 #include <torch/nn/modules/batchnorm.h>
 #include <torch/nn/modules/conv.h>
 #include <torch/nn/modules/dropout.h>
@@ -10,8 +9,8 @@
 #include <torch/nn/modules/functional.h>
 #include <torch/nn/modules/linear.h>
 #include <torch/nn/modules/loss.h>
-#include <torch/nn/modules/maxpool.h>
 #include <torch/nn/modules/modulelist.h>
 #include <torch/nn/modules/named_any.h>
+#include <torch/nn/modules/pooling.h>
 #include <torch/nn/modules/rnn.h>
 #include <torch/nn/modules/sequential.h>
