+++ torch/csrc/api/include/torch/nn/options.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <torch/nn/options/pooling.h>
